requirements.txt
streamlit
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
python-dotenv
